Update patch set 2
Patch Set 2:

I have added cases with "GET" methods, Not sure I should also add cases of "PATCH" and "PUT" methods as it can change the server settings and may break it.

For testing, It is better to disable Testapi (by commenting the lines 573, 574) as after connecting to the VPN your Internet will not work.

Patch-set: 2